requirements.txt
discord.py
colorama
dotenv
requests
PyYAML
pillow
python-dotenv
twitchAPI